+++ backend/apps/accounts/admin.py
@@ -1,5 +1,5 @@
 from django.contrib import admin
 
-from .models import User
+from .user import User
 
 admin.site.register(User)
