+++ backend/apps/accounts/generate_users.py
@@ -9,7 +9,7 @@ from django.conf import settings
 from django.http import HttpResponse
 import numpy as np
 
-from apps.accounts.models import User
+from apps.accounts.user import User
 from apps.iamstudent.models import AUSBILDUNGS_TYPEN_COLUMNS, Student
 from apps.ineedstudent.models import Hospital
