+++ backend/apps/accounts/views.py
@@ -27,8 +27,10 @@ from apps.ineedstudent.views import ApprovalHospitalTable
 
 from .decorator import hospital_required, student_required
 from .forms import CustomAuthenticationForm, NewsletterEditForm, NewsletterViewForm, TestMailForm
-from .models import LetterApprovedBy, Newsletter, NewsletterState, User
+from .newsletter import Newsletter, NewsletterState
+from .newsletter_approved_by import LetterApprovedBy
 from .tables import NewsletterTable
+from .user import User
 
 logger = logging.getLogger(__name__)
